backend/requirements.txt
fastapi==2.0.1
minio==7.1.0
sentry-sdk==1.5.1
nvidia-pyindex==1.0.6
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
databases==0.6.1
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.12.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
MarkupSafe==2.1.1
orjson==3.7.12
psycopg2==2.9.3
py-bcrypt==0.4
pydantic==1.9.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
watchfiles==0.16.1
watchgod==0.8.2
websockets==10.3
aio-pika==8.2.0
aioboto3==10.0.0
fastapi-jwt-auth[asymmetric]

